chore: generate CRUD delete procedures (#210)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

We'd like to offer our users obvious mutations for free. We start by
offering `delete_<tablename>_by_<some_unique_column>`.

`some_unique_column` must be a non-compound key (ie, `int` is fine). If
they key is complex, we don't generate a procedure for it.

### How

We don't store the generated procedures in metadata currently, they are
generated on the fly using `metadata.tables`. I don't think this is
ideal but would rather make sure we're happy with the broad idea here
before getting into that.

We generate them a) when the `/schema` endpoint is called, so we can
tell the user what is available and b) when the user tries to run a
procedure.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>