support queries with no fields specified (#214)
### What

We want to support queries with no fields so we can run queries such as:

```graphql
query {
  Albums(where: ..., limit: 10) {
    __typename
  }
}
```

### How

- We generate a query for rows without bailing even if the query does
not contain fields (which generate a `SELECT FROM ...` query), but we
will also indicate whether we got fields or not (None or an empty
indexmap means no, otherwise yes).
- If the query specifies aggregates and we did not get row fields, we
return the aggregate results fields only.
- If there are no aggregate fields specified, the query will always
return rows result, which might be empty fields.
- We remove the NoFields error option.