Introspection improvements (#221)
### What

This PR improves the introspection query by filtering out types that are
mostly used for server administration, introspection, or other internal
purposes for postgres.

It also improves the function selection heuristic for implicit casts,
preferring alternatives that use fewer casts. It also filters those
casts which are unlikely to be useful.

A nice consequence of this is that the deployment metadata we generate
(for Chinook for instance) is more concise. Specifically we can see that
some of citus's metadata tables disappear from view.