test that failed mutations are rolled back (#237)
### What

We want to test that mutation requests are run as a transaction, and
that if one mutation fails, the other mutations are rolled back.

### How

We add a test with a mutation request that tries to insert an album with
an already existing id, which fails. We check that it fails, and we also
verify that the artist we inserted before was not inserted to the db.