Return the generated SQL of an explain request with empty variables (#241)
### What

To support use cases such as explaining remote-relationships, we want to
be able to view the generated SQL of a query that uses variables but
does not provide variables (since those would need to be run first in
order to be provided).
In this PR we allow explain requests on queries that provide empty
variables (`variables: []`) to receive just the generated SQL, without
the postgres explain part.

### How

We check that we receive an empty variables field and if that's the case
we will skip querying postgres and just return the generated SQL.

Please read this PR with the `hide whitespace` setting.