Support composite types: models and translation (#240)
### What

This PR adds limited support for composite types.

The ndc spec does not provide any means to interact with a
composite/object type other than allowing them to appear in results and
as input arguments. It does not e.g. let you perform field selection or
comparison operators. This PR supports that.

What is not yet supported is the introspection of composite types from
the database. This will appear in a follow-up PR.

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->