introspect columns to include default, generated and identity columns (#256)
### What

In order to generate insert mutations we'd like to know which columns
should be inserted into, which we shouldn't insert to, and which can go
either way.
To do that we add settings for default, generated, and identity
information to column information, as well as introspection code.

### How

1. We create an enum for each setting, and include that information in
`ColumnInfo` metadata
2. We add SQL introspection and query this information from
`pg_attribute`, but see the comment about generated columns