Support variables of composite type (#249)
### What

This PR changes how the `variables` field of a request is translated,
such that variables of composite types are now supported.

Array types require a some treatment of their own, which will be dealt
with in a follow-up PR.

### How

Prior to this PR, incoming variables (bound to `$1` in this example)
were translated in the style of:

```
  ... FROM json_to_recordset($1) as "%variables_table"("%variable_order" int, "variable1" varchar, "variable2" varchar, "variable3" varchar, ...)
```

All the variables used in the request (`variable1`, `variable2`, etc)
all become columns of `%variables_table`, and all of type `varchar`.

Usage sites would rely on casts to interpret the variables as the
relevant scalar type:

```
 cast("%variables_table"."variable1" as int4)
```

On this approach we can observe a few things:

* We can only support plain scalar types, since we cannot canonically
represent an array or object as a string through `json_to_recordset`.
* "%variable_order" is a reserved name that can conflict with user
variable names.
* We don't need to do any bookkeeping of what type a variable has. Usage
context takes care of this to the extent that queries without type
errors execute correctly.

In order to support composite (and array) types, we change the
translation scheme:

```
  ... FROM json_to_recordset($1) as "%variables_table"("%variable_order" int, "%variables" jsonb)
```

All the variables used in the request (`variable1`, `variable2`, etc)
all become fields of the "%variables" column, which remains yet a json
value.

Usage sites still rely on casts to interpret the variables as the
relevant scalar type, but now have to rely on the `->>` operator to
project a variable from the `%variables` object as a plain `text`
scalar:

```
 cast("%variables_table"."%variables" ->> "variable1" as int4)
```

In the case of composite types we use the function
`jsonb_populate_record` to construct the composite value from the
json-valued variable-field, and the operator `->` to extract the field
value as `jsonb` (c.f. `->>` for `text`)

```
jsonb_populate_record(null::person, "%variables_table"."%variables" -> "variable2")
```

On this updated approach we can observe a few things:

* We can support both scalars, arrays, and composite types.
* "%variable_order" is no longer a reserved name.
* We still don't need to do any bookkeeping of what type a variable has.
* `jsonb_populate_record` does all the heavy lifting, which means that
we don't need to generate more than a single call to this function to
handle any level of nested types.

---------

Co-authored-by: Samir Talwar <samir.talwar@hasura.io>