Add tests of composite types to Aurora (#259)
### What

This PR extends the tests of composite types to cover Aurora.

This is a follow-up to #257, and will be rebased once that is merged.

### How

In order to standardise the (rare) management we need to be able to do
on our Aurora test instance this PR adds a script that inserts all the
test data currently in use. Currently, this means: The chinook dataset,
and definitions of composite types.

---------

Co-authored-by: Samir Talwar <samir.talwar@hasura.io>