** What **
This PR improves the diff that we get when checking the idempotency of
the 'configure' action.

** How **

We split the test in two:

One test checks that the re-serializing the persisted/expected
deployment is strictly idempotent. This still produces an awkward diff,
because the ordering of serialized object keys of `serde_json::Value`s
appear to be non-deterministic.

Another test checks that the running the 'configure' action on the
persisted deployment produces an identical deployment by comparing
their pretty-printed json representations. Since these are results of
`RawConfiguration`s being serialized, the key ordering is preserved.