Support variables of array type (#263)
### What

This PR extends the support for query variables to include arrays.

### How

We need to go a bit out of our way to be able to translate JSON arrays
to Postgres arrays, since there is no single builtin function that does
this.

As an example taken from the tests, a variable of type
array-of-`organization` gets translated to the below SQL:

```
(
  SELECT
    array_agg(
      jsonb_populate_record(cast(null as organization), "%0_arr"."elem")
    ) AS "elem"
  FROM
    jsonb_array_elements(("%0_%variables_table"."%variables" -> $2)) AS "%0_arr"("elem")
)

```

Which, being a single-element set-returning expression, can feature both
in a `FROM` clause and in a select-list or other expression.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>