Unify code path for json value translation with variables (#267)
### What

Rather than having two different ways to translate values (depending on
whether they are literals or variables) we instead use the same method
in both cases.

### How

As a consequence we need to have the `State` available in literal value
translation as well as in variable translation, which requires some
follow-up changes in the various other translation functions.

We also duplicate the test of nested array types from the variables-case
to the literal values-case.