Only get types and procedures from unqualified schemas (#271)
### What

In order to be able to unambiguously refer to types and procedures (i.e
to avoid an issue like
https://github.com/hasura/graphql-engine/issues/10100) we want to be
sure that we track the schema types and procedures are defined in,
similar to how we do tables.

This PR implements a minimal first step on the way to this, by simply
only tracking types and procedures defined in schemas appearing in a new
`unqualifiedSchemasForTypesAndProcedures` configuration option.

It is then up to a follow-up PR to implement schema name prefixing.

(Small aside: As an auxiliary improvement, thi PR introduces the use of
the `regclass` etc. types to refer to catalog table entries instead of
hardcoding oids. Incidentally this fixes fetching of table and column
comments on Cockroach.)

### How

We add a new configuration option field
`unqualifiedSchemasForTypesAndProcedures` that defaults to
`["public","pg_catalog","tiger"]`. During introspection we only look in
these schemas for types and procedures.

For consistency we also rename `unqualifiedSchemas` to
`unqualifiedSchemasForTables`. In order to make this change conservative
we still recognize `unqualifiedSchemas` in the API schema and json
parsers, but we will only ever output `unqualifiedSchemasForTables` from
the configuration server.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>