Nix: Fix the build inputs for the shell and app. (#285)
This makes a few changes to the Nix flake to get things building on my
machine.

1. On macOS, add `CoreFoundation` to the list of dependencies.
2. Adds `strictDeps = true` to the package, which means we need to add
`openssl.dev` (and `pkg-config`) to `nativeBuildInputs` as they're
required at build-time.
3. Move `openssl` and `protobuf` to `buildInputs` as they're required at
runtime.
4. Reuse inputs from the default package in the shell.
5. Hide the `perf` shell from macOS so `nix flake check` works.

The last point is partially undoing a change I made earlier, where the
shell brought in inputs from all flake checks. This was way too broad;
this narrows it a lot.