Break dependencies through copying and pasting. (#292)
### What

`ndc-postgres-configuration` had a dependency on `query-engine-sql`
simply to use the `IsolationLevel` enum. This dependency is otherwise
unnecessary and quite large.

### How

I have copied this enum into the configuration so that we can avoid this
large dependency.

Of course, now we have two, so I've added a function inside the
connector crate to convert from one to the other.

I have also improved the OpenAPI documentation on the `IsolationLevel`
enum somewhat.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>