Borrow more when constructing `RuntimeConfiguration`. (#293)
### What

The documentation for `RuntimeConfiguration` states that it borrows
whenever possible, which was false. I have made some changes to borrow
(or copy, if copying is cheap) all fields.

### How

I introduced the `'request` lifetime and annotated fields with it. The
metadata might be borrowed, or it might be converted from a v1 metadata,
so this is wrapped in a `Cow` to abstract over this.

When possible, we now pass around a `RuntimeConfiguration` by value
rather than by reference.

I also added `Copy` to `MutationsVersion` so that we can freely pass it
around without having to worry about borrowing. `IsolationLevel` already
implements `Copy`.