Replace configuration with a new version, identical to v2. (#295)
### What

This introduces configuration version 3, and removes versions 1 and 2.

This is most definitely a breaking change, and we're not done. Future
changes will tear this apart and rebuild it, but I wanted to make this
part easier to review.

### How

I did this in a few steps to make sure I didn't break anything.

1. I pulled out some of the lower-level configuration types to their own
files.
2. Next, I copied the remaining shared code from v1 to v2 so there were
no cross-dependencies.
3. Then I copied v2 to v3, and modified all the tests to use a v3 config
(identical to v2 except the `"version"` field).
4. Finally, I removed v1 and v2.