Bring ndc-postgres in line with ndc-spec RC15 (#296)
## What

This PR updates `ndc-postgres` to adhere to the RC15 version of
`ndc-spec`.

## How

This required a bit more work than we'd originally expected: because the
format of procedure results changed, we now need to generate a new
object type for each procedure (namely, the result type: a product of
`affected_rows` (int4) and `returning` (whatever the underlying
collection's row type is).