Enable Clippy warnings on all targets and features, and fix them. (#300)
### What

This changes our scripts to run Clippy on all targets and features.

It catches a few issues, which I have fixed.

1. Removing unnecessarily-nested modules.
2. Adding a semicolon or two.
3. Using `is_empty`.
4. Using an array instead of a vector for a static list.

I also found out that it's enough to set `RUSTFLAGS="-D warnings"` to
make Clippy fail on warnings, so I have changed CI to rely on that. This
means we don't need to duplicate the list of warnings.

### How

Most of this was covered by adding `--all-targets --all-features` and
running `just lint-fix`.

The rest was deleting a bit of code.

I suggest ignoring whitespace when reviewing.