benchmarks: Generate configuration without the configuration server.
The configuration server is going away. Eventually we'll have a CLI
that does the same thing, but that doesn't currently exist.

Fortunately, specifically for the benchmarks, we can replicate the
configuration server behavior simply by using `jq` to replace the
connection URI in the already-generated configuration file.

I've made the changes in both the local benchmark script and in the CI
YAML file.