Expose equal and in operators as part of the schema response (#304)
### What

rc14 requires we expose the `equal` and `in` operators as part of the
schema response.
To do that, we have to specify them as part of the metadata.

### How

- We add a new attribute to operators - OperatorKind, which will say
whether an operator is equal, in or custom, similar to
[ComparisonOperatorDefinition](https://hasura.github.io/ndc-spec/reference/types.html#comparisonoperatordefinition).
- In the introspection, we map the `_eq` operator to the kind `Equal`,
and other operators to `Custom`.
- In the introspection, we check if a scalar type has the `equal`
operator, and add it the `in` operator as well if yes.
- We change the filtering code so that it checks the operator kind of an
operator to distinguish `in` from the rest, instead of looking at the
name.
- We regenerate the ndc-metadata and fix the tests

We also:

- Moved version3.* into the version3/ dir
- Removed the old `v1` and `v2` config tests.