Switch to a configuration directory. (#305)
### What

This upgrades ndc-sdk, which includes two major changes:

1. Configuration is now a directory, not a `RawConfiguration`.
2. The configuration server has gone.

### How

This means we have to make a lot of changes. I'll try and summarize
them.

1. We now expect the configuration file to be a file inside the
configuration directory. Test data and tests have been changed
accordingly.
2. The configuration directory is assumed to be `/etc/connector` unless
otherwise specified when run as a Docker image.
3. The configuration server has been removed. This means that our
scripts that generate configuration no longer work. We will need to
replace that logic with a CLI or similar.
4. We don't need to deploy the configuration server any more, because it
doesn't exist.