fix just dev workflow (#308)
### What

Fix the `just dev` workflow stuck in an infinite loop.

### How

Mutation tests create a new ndc metadata directory for each test.
Because the directory were not ignored by git, only the config files,
they were picked up as changed to be watched for. We fix the gitignore
and that fix the test.