Remove duplication in error types and error-handling. (#314)
### What

I started trying to re-use the transaction handling logic we have around
mutations for queries too, but got stuck because I either needed to
duplicate a lot of code, which felt suspect and error-prone, or
deduplicate a lot of other code first.

I chose the latter.

This deduplicates the error types between queries and mutations in a new
`query_engine_execution::error::Error` type.

I have also deduplicated the use of this type and the
`query_engine_translation::translation::error::Error` type, particularly
when it comes to (a) recording the error for tracing and metrics, and
(b) converting the error to an ndc-sdk `connector::QueryError`.

I removed the `Multiple` error case, instead deciding to just throw away
the second error from rolling back in favor of the first one, which is
more likely to contain useful information.

I think the separation of concerns in the resulting code makes it easier
to read (large `.map_err` blocks can be something of a handful). This
will also make it much easier to share transaction code between queries
and mutations.

### How

First, I made the new error type. This meant I had to handle some extra
cases in certain places, and I didn't like adding the code, so I decided
instead to remove it.

I created two new modules:

1. `ndc_postgres::error::record`, which records errors with the tracing
and metrics services.
2. `ndc_postgres::error::convert`, which converts errors from internal
query engine types to ndc-sdk types.

From there it was a fairly mechanical process of using those new
functions instead of the inline recording and conversion from before.

I also removed a bit of code:

* `ndc_postgres::query::plan_query` was duplicated in
`ndc_postgres::query::explain::plan_query`, with different
error-handling. By moving error-handling up a layer, the discrepancy
went away and I was able to reuse the first one.
* `ndc_postgres::mutation::explain` had an additional error conversion
function which converted the ndc-sdk error _back_ to an internal error.
Again, by moving error-handling up a layer, this was no longer
necessary.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>