move the ArrayType option lower in the schema (#318)
### What

In the enum `Type` we place `ArrayType(Type)`, a recursive type, first
in the list of options.
Some tooling such as json-editor, don't seem to like that very much, and
would prefer a non-recursive option first.

### How

In the enum definition, we push `ArrayType` down, and that seems to make
the tools happy.