Roll back queries on failure. (#316)
### What

We need to end transactions if they fail. This ensures we do in the case
of queries.

Mutations were already handled.

### How

I have moved the `rollback_on_exception` function to its own module and
ensured that it is used in `query_engine_execution::query`.

To test this, I have constructed new metadata with just a couple of
native queries: one broken, one working. The broken one is run first,
then the working one afterwards. Before this change, the test failed
because the working query would fail with the error:

```
error returned from database: current transaction is aborted, commands ignored until end of transaction block
```

### Limitations

Note that there are a couple of limitations with this implementation:

1. This behavior is optional. Adding more entrypoints into query
execution will require us to remember to correctly handle transactions
there too.
2. On a panic, the transaction will not be rolled back.

In the future, we may want to consider:

1. Delegating this behavior to `sqlx`. (We tried this but it meant an
extra query to compensate for the `begin` function not taking
transaction options, which degrades performance.)
2. Wrapping all database access to ensure we handle transactions
correctly.
3. Handling panics.