Break the dependency on ndc-sdk in the configuration crate. (#331)
### What

By maintaining our own error type, we can avoid this dependency, which
means the CLI also doesn't need to bring in ndc-sdk.

This means that the CLI does not need to depend on OpenSSL.

### How

I copied the `ParseError` enum, with a few variations, into the
configuration crate, and used that instead.

The connector crate now maps from one to the other.