Publish CLI binaries for all target platforms. (#332)
### What

In order for the Hasura CLI to work with ndc-postgres, it needs to know
about our CLI.

Which means we need to build it.

For now, we just store it in GitHub artifacts. Later, we will make this
part of the release process, and releases will have binaries attached to
them.

### How

I have added a GitHub Actions job to build the CLI and publish an
artifact.

It does this for 5 platforms:

- Linux on x86_64 (`x86_64-unknown-linux-gnu`)
- Linux on aarch64 (`aarch64-unknown-linux-gnu`)
- macOS on x86_64 (`x86_64-apple-darwin`)
- macOS on aarch64 (`aarch64-apple-darwin`)
- Windows on x86_64 (`x86_64-pc-windows-msvc`)