Optionally initialize the configuration directory with metadata. (#333)
### What

To publish the connector to the NDC Hub, we need to give it some
metadata, containing information such as the Docker image + tag, the CLI
plugin name, etc.

This information is specified in the [Packaging
RFC](https://github.com/hasura/ndc-hub/blob/main/rfcs/0001-packaging.md#connector-definition).

I've added a `--with-metadata` flag to the CLI `initialize` command
which creates this metadata. It's unlikely a user will need this, but it
feels like the best place for it, and there is no harm in allowing a
user to do this themselves.

### How

I've converted the metadata schema into Rust types. We construct the
appropriate metadata, serialize it to YAML, and write it to disk in the
correct location.

We capture the `RELEASE_VERSION` environment variable at build time, and
if it's set, use that for versions. Otherwise we use `latest`. (I've
told `clap` about this too so you can run `ndc-postgres-cli --version`.)

In release builds, we capture Git information if `RELEASE_VERSION` is
not set; in debug builds, we don't bother.