Add the CLI binary name to the connector manifest. (#344)
### What

This is required for the CLI to work. The binary name chosen is the one
in the [CLI plugins index][CLI plugin manifest].

This couples the connector manifest to the CLI plugin manifest, which
isn't ideal. We may re-negotiate this at a future time, e.g. by adding
syntax to this file to instruct the CLI to prefix the binary.

[CLI plugin manifest]:
https://github.com/hasura/cli-plugins-index/blob/master/plugins/ndc-postgres/v0.4.0/manifest.yaml

### How

Typing.