Rename schema fields and add a jsonschema generation and ref on init (#361)
### What

This PR does the following:

1. It renames a few fields in the configuration:
   - `configureOptions` -> `introspectionOptions`
   - `mutationsVersion` moved from `configureOptions` to the top-level
- connection uri, pool settings and isolation level are now nested under
"connectionSettings"
- Native queries use a new `ReadonlyColumnInfo` type instead of
`ColumnInfo` type for arguments and column. ReadonlyColumnInfo does not
contain default, identity and generated information which we don't want
to appear in autocomplete for native queries.
2. It generates a `schema.jsonschema` file on cli `initialize` command
next to the `configuration.json` file
3. It adds the `$schema` field to the configuration, referencing the
generated `schema.jsonschema` file.

The purpose of (2) and (3) is to provide auto-complete features when
editing the configuration in vscode (demo soon).

![Screenshot from 2024-03-12
16-56-07](https://github.com/hasura/ndc-postgres/assets/8547573/f4149d91-041a-4571-b387-0f873cc9ad21)