update command should not override the config file in some cases (#362)
### What

We want the update command in ndc-postgres-cli to not override changed
made after we started introspecting but before we wrote the changes to
file, and not to write to the filesystem when the file does not need
changing.

### How

1. We loop the update operation 3 times
2. We read the input file after the introspection took place as well
3. If the input from before the introspection and the input from after
the introspection are the same we continue, otherwise we try again
4. Before writing to file, we check that if the output is going to be
the same as the input, if it is the same, we skip writing, if not, we
write.
5. We add the `Eq` trait everywhere so we can compare the config.