fix names of fields in schema generation (#366)
### What

The `/schema` endpoint was exposing the wrong column names in object
types. This PR fixes this.

### How

columns, referenced by tables, native queries, etc. Have a type that
looks somewhat like this: `BTreeMap<String, ColumnInfo>`.

`ColumnInfo` looks somewhat like this: `ColumnInfo { name: String, ...
}`.

The `String` in the map refers to the external name that ndc-spec
queries should reference, while the `ColumnInfo` name refers to the
database column name.

By mistake, were reporting the `ColumnInfo` name in the schema, instead
of the String from the map.

We fix this by referring to the right name, and use `iter()` on the
BTreeMap instead of `values()`.