Build the CLI on Ubuntu 20.04. (#373)
### What

Downgrading Ubuntu from "latest" (22.04) to 20.04 means that it will
also be supported on that version and all subsequent versions. Currently
it fails to start on Ubuntu 20.04 because the version of glibc linked
against is too new.

### How

I have downgraded the Ubuntu runner we use for building the CLI from
22.04 to 20.04.

I have also fixed the Rust build cache to take the runner into account.