Read Native Query SQL from files (#372)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

Currently we can specify native queries by inlining their SQL as a
string in the `sql` field.
We'd like to be able to reference a `.sql` file instead, so we can avoid
escaping characters and can use newlines.

The new config format will accept the following structures:

```json
"sql": "<inline sql>" (as before, but also:)

"sql": { "inline": "<inline sql>" }

"sql": { "file": "<relative file path>" }
```

### How

- We parse the `NativeQuerySql` which represents the SQL field using an
intermediate new type `NativeQuerySqlExternal`.
- as part of the parse_configuration step and after serde
deserialization, we traverse the metadata and convert
`NativeQuerySqlExternal` to `NativeQuerySql` by reading from file
(performing IO) if needed, and parsing to `NativeQueryParts`. We use the
configuration directory that the user supplied to locate the relative
path.

We replace the existing `sql` field in nativequeryinfo with an Either
external or internal representation. This is not ideal because it breaks
the parse don't validate principle.
We would like to mitigate this in the future by separating the
`RawConfiguration` types and the (execution time) `Configuration` types
completely.
We will do that as part of a larger ticket of redesigning the
configuration format so it can be split across multiple files.