Add support for introspecting domain types (#380)
### What

This PR adds support for introspecting domain types.

### How

Prior to this PR domain types were already recognised as scalar types.
However, the ability to view a domain type transparently as its base
type was not, so you couldn't _do_ anything with values of domain type.

This PR adds cases to the set of implicit casts for domain types in the
introspection query, which enables support for comparison operators.