Support enum types (#387)
### What

This PR improves introspection support for enum types.

Comparison operators and aggregation functions on enum types are
implicitly defined in PostgreSQL and thus were not detected prior to
this PR.

### How

We simply collect which enum types are defined and make up the
conventional comparison operators and the min/max aggregation functions
for all enum types.

We test that we can select and aggregate enum types, and that we can
insert enum values which PostgreSQL validates.

We also collect the enum value labels, which we will be using once
ndc-spec v0.1.1 is released.