Upgrade ndc-spec to v0.1.1.
See https://github.com/hasura/ndc-sdk-rs/pull/1 for details.

I discovered we can specify workspace dependencies, which means the
versions only need to go in one place. I am experimenting with this
Cargo feature in this changeset. If we like it, we may want to use it
for all dependencies.