Upgrade ndc-spec to v0.1.1. (#390)
### What

This includes:

- type representations in the models
- upgrading the OpenTelemetry SDK
- adding a feature switch to support `rustls` instead of `native-tls`
- a lot of enhancements to ndc-test

ndc-sdk was upgraded in https://github.com/hasura/ndc-sdk-rs/pull/1.

We do not actually specify type representations in this changeset. We
will add these in a follow-up change.

### How

I discovered we can specify workspace dependencies, which means the
versions only need to go in one place. I am experimenting with this
Cargo feature in this changeset. If we like it, we may want to use it
for all dependencies.