experimental mutations version with boolean expressions (#393)
### What

We introduce a new experimental mutations version. This version is
intended to be used as a WIP version not for public consumption.
It also introduces an update delete mutation that contains a `filter`
predicate field.

<!-- Consider: do we need to add a changelog entry? -->

### How

1. Introduce a new mutations version: `VeryExperimentalWip`.
2. Take the existing `insert`, `delete`, `translate` and `generate`
modules, duplicate them, and place them in `v1/` and `experimental/`
directories.
3. Add and modify existing code to dispatch on operations from the right
module according to the mutations version in the config.
4. Add a new `filter` field for the delete operation which will be
interpreted as a boolean expression and added as a `WHERE` clause in the
delete sql.
5. Expose this new field as part of the schema as a boolean expression
over the collection.